Added image support from chapter 7 of McGaw
Pulled out the image support I had before, added in his stuff with changes to match my version of Django.  Not sure if the image caption will stay, but it's here regardless.  Also noted two problems in product.djhtml -- ugly formatted content and it's not displaying the value for p.abv, why?