Fixed #8.
According to #8, inactive jars were still being counted for purposes of calculating crate capacity.  A test was written to check this bug.  The JSON was modified to contain additional inactive jars.  The test failed -- which was expected.  The code in the model was changed until the test passed.  This is also good.