Successfully migrated from SQLite3 to Postgres.  Fixes #12.
Most of the conversion was in the database, not in the code.  The only
code changes were configuring the default settings to point to
Postgres and to add the necessary packages to the requirements.txt.
The changes in README.md are just reformatting, no actual content
changes, I was just annoyed at the word wrap!