As part of release 0.8 I am changing from 'alpha' to 'beta'.
The application *as is* is now something I am using for production.
There will be additional major features -- the meadery app and the
recipe app -- but those features are not necessary for me to use the
application at this time.  I will note those features as alpha and
beta as appropriate when they are added.