Fixed product ordering and added inventory drilldown.  Fixes #2.
The product ordering has been changed to push inactive products to the
bottom and sort from newest to oldest.  The link approach used for
categories has been applied to inventory and provides a drill-down
from the warehouse level all the way to the crate.

This completes all of #2 that can be fixed without the ability to
properly set a default filter.