Changes and additions to models, plus tests.  Refs #18.
Honey changed to kilograms, and methods were added for converting
between mass and volume.  Changed recipe references from "amounts" to
masses and volumes, changed both to use Decimal instead of Integer and
Float.  Added methods to Recipe to predict brew temperature and
specific gravity as well as the beginnings of natural and appellation
checks and added SG correction to samples.

Still haven't added support for sets of ingredients (such as two
different types of honey or flavors) or volume projections for
recipes.