Changed slugs for inventory models, improved tests.
The major change here was to the inventory model's slug definition.
This will be handy when inventory views are written.  The inventory
fixtures were updated to support additional tests.  The management
commands were modified to require proper options.  The inventory tests
were completed, with the exception of one management command which
will likely be converted to an inventory view.