Cleaned fixtures and tests, fixed bug in meadery, added correctness tests!
There was a bug in the meadery method that created products from
batches -- when the same batch was used to create a product more than
once, a server error was returned.  The test caught the bug.  I fixed
the bug.  The test passes without errors.  Victory!

Correctness tests have been added to IngredientTestCase.  The add test
checks that the ingredient did not exist before the test, and that it
did exist after the test.  The modify test checks that the attribute
in question was modified after the test.  The delete test checks that
the ingredient was in fact deleted.  Similar changes will be made to
other tests.  The hard part will be adding correctness tests to less
simple situations.