Finished accounts testing, for now.
Updated accounts fixture to include actual accounts stuff, not just
auth.User -- and why is email in both?  Anyway!  Modified checkout
fixture to reference accounts fixture to make testing easier.  Found a
minor bug in the urls.py that goes back to McGaw, and of course wrote
a bunch of tests.