Bring Docker support up-to-date
The state of the art in Docker containers has advanced quite a bit since
support was originally added.

The container image has reverted to Python 2.7, and the squid-deb-proxy support
has been removed.  The data-only containers have been replaced with named
volumes and moved to the root directory.  The app itself now resides in /code.

The dev setup script has also been cleaned up quite a bit.  A number of ugly
workarounds are no longer necessary and were removed.

The Dockerfile has been updated to version 2 and cleaned up in general.

The worst kludge of them all was replaced with a best-practices script which
waits for the database to start before performing db-related app tasks.  Many
thanks to https://github.com/vishnubob/wait-for-it for solving that problem!