Documentation update
The requirements were updated in the README to reflect the new dependence on
recent versions of Docker Engine and Docker Compose.  The development info
was removed and minor typos were addressed.

A comment in dev-setup.sh was fixed.

The docker-compose.yml file required a minor change to add the --insecure flag
for the web server.

The dev documentation was updated to reflect changes in docker-compose and to
mention the documentation itself, plus the year bump on the copyright page.

The requirements.txt file had excess baggage which is no longer required.